Feat : 신고하기 취소 함수를 commentModal 로 옮김
